frontend/pages/AskElliot.py
import streamlit as st
import requests
import os
from dotenv import load_dotenv

load_dotenv()

API_KEY = os.getenv("MISTRAL_API_KEY")
API_URL = "https://api.mistral.ai/v1/chat/completions"

st.title("Ask Elliot (Mr. Robot)")

# Initialize chat history
if "messages" not in st.session_state:
    st.session_state.messages = []

# Display chat history
for message in st.session_state.messages:
    with st.chat_message(message["role"]):
        st.write(message["content"])

# Chat input box
user_input = st.chat_input("Type your message...")

if user_input:
    # Add user message
    st.session_state.messages.append({"role": "user", "content": user_input})

    with st.chat_message("user"):
        st.write(user_input)

    headers = {
        "Authorization": f"Bearer {API_KEY}",
        "Content-Type": "application/json",
    }

    payload = {"model": "mistral-small-latest", "messages": st.session_state.messages}

    response = requests.post(
        API_URL,
        json=payload,
        headers=headers,
        timeout=30,
    )

    if response.status_code == 200:
        answer = response.json()["choices"][0]["message"]["content"]

        # Add assistant message
        st.session_state.messages.append({"role": "assistant", "content": answer})

        with st.chat_message("assistant"):
            st.write(answer)

    else:
        st.error(response.text)
