pyproject.toml
[project]
name = "pyintro"
version = "0.1.0"
description = "CVE Dashboard for Intro to Python Course"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.136.0",
    "matplotlib>=3.10.8",
    "networkx>=3.6.1",
    "pandas>=3.0.2",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "streamlit>=1.56.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pyright>=1.1.408",
    "pytest>=9.0.3",
    "ruff>=0.15.11",
]
